Revert "Protected monolog (#5949)"
This reverts commit 7ff6ad04a6ad1b9c86bc032165d0a131eda9582a.
