+++ composer.json
@@ -42,7 +42,8 @@
 	"require": {
 		"php": ">=7.3",
 		"cloudflare/cf-ip-rewrite": "^1.0",
-		"composer/installers": "^1.0 || ^2.0"
+		"composer/installers": "^1.0 || ^2.0",
+		"monolog/monolog": "^1.0 || ^2.0"
 	},
 	"require-dev": {
 		"php": "^7 || ^8",
@@ -72,8 +73,7 @@
 		"wpackagist-plugin/simple-custom-css": "^4.0.3",
 		"wpackagist-plugin/spinupwp": "^1.1",
 		"wpackagist-plugin/woocommerce": "^7",
-		"wpackagist-plugin/wp-smushit": "^3",
-		"wp-media/monolog": "^0.0"
+		"wpackagist-plugin/wp-smushit": "^3"
 	},
 	"autoload": {
 		"classmap": [
@@ -110,7 +110,6 @@
 				"wp-media/background-processing",
 				"wp-media/rocket-lazyload-common",
 				"wp-media/wp-imagify-partner",
-				"wp-media/monolog",
 				"league/container"
 			]
 		}
