--- inc/Dependencies/Monolog/Formatter/NormalizerFormatter.php
@@ -1,199 +1,0 @@
-<?php
-
-/*
- * This file is part of the WP_Rocket\Dependencies\Monolog package.
- *
- * (c) Jordi Boggiano <j.boggiano@seld.be>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace WP_Rocket\Dependencies\Monolog\Formatter;
-
-use Exception;
-use WP_Rocket\Dependencies\Monolog\Utils;
-
-/**
- * Normalizes incoming records to remove objects/resources so it's easier to dump to various targets
- *
- * @author Jordi Boggiano <j.boggiano@seld.be>
- */
-class NormalizerFormatter implements FormatterInterface
-{
-    const SIMPLE_DATE = "Y-m-d H:i:s";
-
-    protected $dateFormat;
-    protected $maxDepth;
-
-    /**
-     * @param string $dateFormat The format of the timestamp: one supported by DateTime::format
-     * @param int $maxDepth
-     */
-    public function __construct($dateFormat = null, $maxDepth = 9)
-    {
-        $this->dateFormat = $dateFormat ?: static::SIMPLE_DATE;
-        $this->maxDepth = $maxDepth;
-        if (!function_exists('json_encode')) {
-            throw new \RuntimeException('PHP\'s json extension is required to use WP_Rocket\Dependencies\Monolog\'s NormalizerFormatter');
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function format(array $record)
-    {
-        return $this->normalize($record);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function formatBatch(array $records)
-    {
-        foreach ($records as $key => $record) {
-            $records[$key] = $this->format($record);
-        }
-
-        return $records;
-    }
-
-    /**
-     * @return int
-     */
-    public function getMaxDepth()
-    {
-        return $this->maxDepth;
-    }
-
-    /**
-     * @param int $maxDepth
-     */
-    public function setMaxDepth($maxDepth)
-    {
-        $this->maxDepth = $maxDepth;
-    }
-
-    protected function normalize($data, $depth = 0)
-    {
-        if ($depth > $this->maxDepth) {
-            return 'Over '.$this->maxDepth.' levels deep, aborting normalization';
-        }
-
-        if (null === $data || is_scalar($data)) {
-            if (is_float($data)) {
-                if (is_infinite($data)) {
-                    return ($data > 0 ? '' : '-') . 'INF';
-                }
-                if (is_nan($data)) {
-                    return 'NaN';
-                }
-            }
-
-            return $data;
-        }
-
-        if (is_array($data)) {
-            $normalized = array();
-
-            $count = 1;
-            foreach ($data as $key => $value) {
-                if ($count++ > 1000) {
-                    $normalized['...'] = 'Over 1000 items ('.count($data).' total), aborting normalization';
-                    break;
-                }
-
-                $normalized[$key] = $this->normalize($value, $depth+1);
-            }
-
-            return $normalized;
-        }
-
-        if ($data instanceof \DateTime) {
-            return $data->format($this->dateFormat);
-        }
-
-        if (is_object($data)) {
-            // TODO 2.0 only check for Throwable
-            if ($data instanceof Exception || (PHP_VERSION_ID > 70000 && $data instanceof \Throwable)) {
-                return $this->normalizeException($data);
-            }
-
-            // non-serializable objects that implement __toString stringified
-            if (method_exists($data, '__toString') && !$data instanceof \JsonSerializable) {
-                $value = $data->__toString();
-            } else {
-                // the rest is json-serialized in some way
-                $value = $this->toJson($data, true);
-            }
-
-            return sprintf("[object] (%s: %s)", Utils::getClass($data), $value);
-        }
-
-        if (is_resource($data)) {
-            return sprintf('[resource] (%s)', get_resource_type($data));
-        }
-
-        return '[unknown('.gettype($data).')]';
-    }
-
-    protected function normalizeException($e)
-    {
-        // TODO 2.0 only check for Throwable
-        if (!$e instanceof Exception && !$e instanceof \Throwable) {
-            throw new \InvalidArgumentException('Exception/Throwable expected, got '.gettype($e).' / '.Utils::getClass($e));
-        }
-
-        $data = array(
-            'class' => Utils::getClass($e),
-            'message' => $e->getMessage(),
-            'code' => (int) $e->getCode(),
-            'file' => $e->getFile().':'.$e->getLine(),
-        );
-
-        if ($e instanceof \SoapFault) {
-            if (isset($e->faultcode)) {
-                $data['faultcode'] = $e->faultcode;
-            }
-
-            if (isset($e->faultactor)) {
-                $data['faultactor'] = $e->faultactor;
-            }
-
-            if (isset($e->detail)) {
-                if  (is_string($e->detail)) {
-                    $data['detail'] = $e->detail;
-                } elseif (is_object($e->detail) || is_array($e->detail)) {
-                    $data['detail'] = $this->toJson($e->detail, true);
-                }
-            }
-        }
-
-        $trace = $e->getTrace();
-        foreach ($trace as $frame) {
-            if (isset($frame['file'])) {
-                $data['trace'][] = $frame['file'].':'.$frame['line'];
-            }
-        }
-
-        if ($previous = $e->getPrevious()) {
-            $data['previous'] = $this->normalizeException($previous);
-        }
-
-        return $data;
-    }
-
-    /**
-     * Return the JSON representation of a value
-     *
-     * @param  mixed             $data
-     * @param  bool              $ignoreErrors
-     * @throws \RuntimeException if encoding fails and errors are not ignored
-     * @return string
-     */
-    protected function toJson($data, $ignoreErrors = false)
-    {
-        return Utils::jsonEncode($data, null, $ignoreErrors);
-    }
-}
