--- inc/Dependencies/Monolog/Handler/ProcessableHandlerInterface.php
@@ -1,40 +1,0 @@
-<?php declare(strict_types=1);
-
-/*
- * This file is part of the WP_Rocket\Dependencies\Monolog package.
- *
- * (c) Jordi Boggiano <j.boggiano@seld.be>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace WP_Rocket\Dependencies\Monolog\Handler;
-
-use WP_Rocket\Dependencies\Monolog\Processor\ProcessorInterface;
-
-/**
- * Interface to describe loggers that have processors
- *
- * This interface is present in monolog 1.x to ease forward compatibility.
- *
- * @author Jordi Boggiano <j.boggiano@seld.be>
- */
-interface ProcessableHandlerInterface
-{
-    /**
-     * Adds a processor in the stack.
-     *
-     * @param  ProcessorInterface|callable $callback
-     * @return HandlerInterface            self
-     */
-    public function pushProcessor($callback): HandlerInterface;
-
-    /**
-     * Removes the processor on top of the stack and returns it.
-     *
-     * @throws \LogicException In case the processor stack is empty
-     * @return callable
-     */
-    public function popProcessor(): callable;
-}
