--- inc/Dependencies/Monolog/Handler/ProcessableHandlerTrait.php
@@ -1,73 +1,0 @@
-<?php declare(strict_types=1);
-
-/*
- * This file is part of the WP_Rocket\Dependencies\Monolog package.
- *
- * (c) Jordi Boggiano <j.boggiano@seld.be>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace WP_Rocket\Dependencies\Monolog\Handler;
-
-use WP_Rocket\Dependencies\Monolog\ResettableInterface;
-
-/**
- * Helper trait for implementing ProcessableInterface
- *
- * This trait is present in monolog 1.x to ease forward compatibility.
- *
- * @author Jordi Boggiano <j.boggiano@seld.be>
- */
-trait ProcessableHandlerTrait
-{
-    /**
-     * @var callable[]
-     */
-    protected $processors = [];
-
-    /**
-     * {@inheritdoc}
-     * @suppress PhanTypeMismatchReturn
-     */
-    public function pushProcessor($callback): HandlerInterface
-    {
-        array_unshift($this->processors, $callback);
-
-        return $this;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function popProcessor(): callable
-    {
-        if (!$this->processors) {
-            throw new \LogicException('You tried to pop from an empty processor stack.');
-        }
-
-        return array_shift($this->processors);
-    }
-
-    /**
-     * Processes a record.
-     */
-    protected function processRecord(array $record): array
-    {
-        foreach ($this->processors as $processor) {
-            $record = $processor($record);
-        }
-
-        return $record;
-    }
-
-    protected function resetProcessors(): void
-    {
-        foreach ($this->processors as $processor) {
-            if ($processor instanceof ResettableInterface) {
-                $processor->reset();
-            }
-        }
-    }
-}
