--- inc/Dependencies/Psr/Log/InvalidArgumentException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace WP_Rocket\Dependencies\Psr\Log;
-
-class InvalidArgumentException extends \InvalidArgumentException
-{
-}
