--- inc/Dependencies/Psr/Log/LoggerAwareInterface.php
@@ -1,18 +1,0 @@
-<?php
-
-namespace WP_Rocket\Dependencies\Psr\Log;
-
-/**
- * Describes a logger-aware instance.
- */
-interface LoggerAwareInterface
-{
-    /**
-     * Sets a logger instance on the object.
-     *
-     * @param LoggerInterface $logger
-     *
-     * @return void
-     */
-    public function setLogger(LoggerInterface $logger);
-}
