--- inc/Dependencies/Psr/Log/NullLogger.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace WP_Rocket\Dependencies\Psr\Log;
-
-/**
- * This Logger can be used to avoid conditional log calls.
- *
- * Logging should always be optional, and if no logger is provided to your
- * library creating a NullLogger instance to have something to throw logs at
- * is a good way to avoid littering your code with `if ($this->logger) { }`
- * blocks.
- */
-class NullLogger extends AbstractLogger
-{
-    /**
-     * Logs with an arbitrary level.
-     *
-     * @param mixed  $level
-     * @param string $message
-     * @param array  $context
-     *
-     * @return void
-     *
-     * @throws \WP_Rocket\Dependencies\Psr\Log\InvalidArgumentException
-     */
-    public function log($level, $message, array $context = array())
-    {
-        // noop
-    }
-}
